Refactoring the system: -  in three header files -two implementation files - and the main file.
