+++ xwingbase/Hardware/Sketches/Lights_Bunker.h
@@ -1,0 +1,4 @@
+// interval definitions in milliseconds
+const unsigned long large_silence_interval = 1000;
+const unsigned long short_silence_interval = 100;
+const unsigned long short_emission_interval = 100;
