+++ xwingbase/Hardware/Sketches/Lights_Droid.h
@@ -1,0 +1,9 @@
+// helper methods  
+private void setupDroidLights(struct DroidLightsParameters*);
+private void setDroidLightStatus(int, struct DroidLightsParameters*);
+
+static const int defaultDuration = 270; // ceil(4 seconds /15 points)
+static const int maxDuration = 4000; // max duration of a cycle in milliseconds
+static const int separationInterval = 100; // a millisecond duration to separate "points" visually
+static const int maxPoints = 15; // max occurrences of a "point" (up signal of varying length)
+static const int durations[3] = {100, 300, 500}; // possible durations of a point (short, medium, long)
